Fix scrolling feature bug
The bug was causing the screen to not show maps near the bottom
and left side of the screen, except those on the corners, for
certain map and screen sizes. The problem was due to not picking
all the neccasry points on the screen that would be used to draw
the maps relative too.