Refactor code
The Map file now contains a scroll function.
This function returns a tupal with two objects
inside of it. The first object is a Screen object. The screen
is what the user will use to 'move' the background
map. The second object is an array of SubMap objects.
Each SubMap object will have a pair of map indicies and
the position of that map's top left corner relative to the
screen's top left corner. This should make it much easier
to work with the scrollable background feature.